fix: cancellationSection was breaking other checkouts
the cancellationSection was being rendered on sections it was not needed
