+++ support-frontend/assets/components/subscriptionCheckouts/summary.module.scss
@@ -211,14 +211,6 @@
   }
 }
 
-.directDebitTerms {
-  a, a:visited {
-    color: gu-colour(neutral-7);
-  }
-  p {
-    margin-top: 10px;
-  }
-}
 .cancellationPolicy {
   display: none;
